+++ includes/abstracts/abstract-wc-heidelpay-payment-gateway.php
@@ -51,15 +51,18 @@ abstract class WC_Heidelpay_Payment_Gateway extends WC_Payment_Gateway
         $this->title = $this->get_option('title');
         $this->description = $this->get_option('description');
         $this->instructions = $this->get_option('instructions');
-        
+
         // Actions
         add_action('woocommerce_update_options_payment_gateways_' . $this->id, array($this, 'process_admin_options'));
         add_action('woocommerce_api_' . strtolower(get_class($this)), array($this, 'callback_handler'));
+        add_action('woocommerce_api_push', array($this, 'pushHandler'));
         add_action('wp_enqueue_scripts', array($this, 'enqueue_assets'));
         add_action('woocommerce_after_checkout_validation', array($this, 'checkoutValidation'));
+        add_action('woocommerce_email_before_order_table', array($this, 'emailInstructions'), 10, 3);
 
         // Filter
         add_filter('woocommerce_available_payment_gateways', array($this, 'setAvailability'));
+        add_filter('woocommerce_thankyou_order_received_text', array($this, 'addPayInfo'));
     }
 
     /**
@@ -67,27 +70,14 @@ abstract class WC_Heidelpay_Payment_Gateway extends WC_Payment_Gateway
      */
     abstract protected function setPayMethod();
 
-    /**
-     * Validate the customer input coming from checkout.
-     * @return boolean
-     */
-    public function checkoutValidation()
-    {
-        //return true;
-    }
 
-    /**
-     * Check whether this paymethod was selected based on
-     * @return bool
-     */
-    public function isGatewayActive()
+    public function pushHandler()
     {
-        if(!empty($_POST['payment_method'])) {
-            if($_POST['payment_method'] === $this->id)
-                return true;
+        if (array_key_exists('<?xml_version', $_POST)) {
+            $push = new WC_Heidelpay_Push();
+            $push->init(file_get_contents('php://input'), $this->get_option('secret'));
         }
-
-        return false;
+        exit;
     }
 
     public function init_form_fields()
@@ -124,7 +114,7 @@ abstract class WC_Heidelpay_Payment_Gateway extends WC_Payment_Gateway
                     'Instructions that will be added to the thank you page and emails.',
                     'woocommerce-heidelpay'
                 ),
-                'default' => __('The following account will be billed:', 'woocommerce-heidelpay'),
+                'default' => '',
                 'desc_tip' => true,
             ),
             'security_sender' => array(
@@ -171,6 +161,29 @@ abstract class WC_Heidelpay_Payment_Gateway extends WC_Payment_Gateway
                 'default' => 'yes'
             ),
         );
+    }
+
+    /**
+     * Validate the customer input coming from checkout.
+     * @return boolean
+     */
+    public function checkoutValidation()
+    {
+        //return true;
+    }
+
+    /**
+     * Check whether this paymethod was selected based on
+     * @return bool
+     */
+    public function isGatewayActive()
+    {
+        if (!empty($_POST['payment_method'])) {
+            if ($_POST['payment_method'] === $this->id)
+                return true;
+        }
+
+        return false;
     }
 
     /**
@@ -280,7 +293,7 @@ abstract class WC_Heidelpay_Payment_Gateway extends WC_Payment_Gateway
         global $wp_version;
 
         $shopType = 'WordPress: ' . $wp_version . ' - ' . 'WooCommerce: ' . wc()->version;
-        $this->payMethod->getRequest()->getCriterion()->set('PUSH_URL', 'push-url for testing'); //TODO insert URL
+        $this->payMethod->getRequest()->getCriterion()->set('PUSH_URL', get_home_url() . '/wc-api/push');
         $this->payMethod->getRequest()->getCriterion()->set('SHOP.TYPE', $shopType);
         $this->payMethod->getRequest()->getCriterion()->set(
             'SHOPMODULE.VERSION',
@@ -318,6 +331,7 @@ abstract class WC_Heidelpay_Payment_Gateway extends WC_Payment_Gateway
                 ];
             }
 
+            $this->paymentLog($this->payMethod->getResponse()->getError());
             $this->addPaymentError($this->getErrorMessage());
         } else {
             $this->addPaymentError($this->getErrorMessage());
@@ -326,7 +340,10 @@ abstract class WC_Heidelpay_Payment_Gateway extends WC_Payment_Gateway
                 WC_Log_Levels::ERROR,
                 htmlspecialchars(
                     print_r(
-                        $this->plugin_id . ' - ' . $this->id . __(' Error: Paymentmethod was not found: ', 'woocommerce-heidelpay') . $this->bookingAction,
+                        $this->plugin_id . ' - ' . $this->id . __(
+                            ' Error: Paymentmethod was not found: ',
+                            'woocommerce-heidelpay'
+                        ) . $this->bookingAction,
                         1
                     )
                 )
@@ -334,6 +351,13 @@ abstract class WC_Heidelpay_Payment_Gateway extends WC_Payment_Gateway
 
             return null;
         }
+    }
+
+    /**
+     * process the Form input from customer comimg from checkout.
+     */
+    protected function handleFormPost()
+    {
     }
 
     /**
@@ -353,13 +377,6 @@ abstract class WC_Heidelpay_Payment_Gateway extends WC_Payment_Gateway
             __('Payment error: ', 'woocommerce-heidelpay') . htmlspecialchars($message),
             'error'
         );
-    }
-
-    /**
-     * process the Form input from customer comimg from checkout.
-     */
-    protected function handleFormPost()
-    {
     }
 
     /**
@@ -401,8 +418,8 @@ abstract class WC_Heidelpay_Payment_Gateway extends WC_Payment_Gateway
      */
     public function callback_handler()
     {
-        $response = new WC_Heidelpay_Response();
         if (!empty($_POST)) {
+            $response = new WC_Heidelpay_Response();
             $response->init($_POST, $this->get_option('secret'));
         }
         exit();
@@ -435,5 +452,86 @@ abstract class WC_Heidelpay_Payment_Gateway extends WC_Payment_Gateway
             'label' => __('Choose a bookingmode', 'woocommerce-heidelpay'),
             'default' => 'DB'
         );
+    }
+
+    /**
+     * Funktion to log Events as a notice. It has a prefix to identify that the log entry is from heidelpay and which
+     * function has created it.
+     * @param  string|array $logData
+     */
+    protected function paymentLog($logData)
+    {
+        $callers = debug_backtrace();
+        wc_get_logger()->log(WC_Log_Levels::NOTICE, print_r('heidelpay - ' .
+            $callers [1] ['function'] .': '. print_r($logData, 1), 1));
+    }
+
+    /**
+     * Get the order using the Get parameter 'key'
+     * @return bool|WC_Order|WC_Refund
+     */
+    public function getOrderFromKey()
+    {
+        if(isset($_GET['key'])) {
+            $order_id = wc_get_order_id_by_order_key($_GET['key']);
+            return wc_get_order($order_id);
+        }
+
+        return null;
+    }
+
+    /**
+     * "woocommerce_thankyou_order_received_text" hook to display heidelpay-paymentInfo text on the successpage after
+     * payment.
+     * @param $orderReceivedText
+     * @return string
+     */
+    public function addPayInfo($orderReceivedText)
+    {
+        $order = $this->getOrderFromKey();
+
+        if ($order === null || $order->get_payment_method() !== $this->id) {
+            return $orderReceivedText;
+        }
+
+        $paymentInfo = $order->get_meta('heidelpay-paymentInfo');
+
+        if(!empty($paymentInfo)) {
+            $orderReceivedText .= '<p>' . $paymentInfo . '</p>';
+        }
+
+        return $orderReceivedText;
+    }
+
+    /**
+     * Hook - "woocommerce_email_before_order_table". Add heidelpay-paymentInfo text to "completed order" email.
+     * @param WC_Order $order
+     * @param $sent_to_admin
+     * @param bool $plain_text
+     * @return null
+     */
+    public function emailInstructions(WC_Order $order, $sent_to_admin, $plain_text = false)
+    {
+        if ($order->get_payment_method() !== $this->id) {
+            return null;
+        }
+
+        if ($this->instructions) {
+            echo wpautop(wptexturize($this->instructions)) . PHP_EOL;
+        }
+
+        $status = $order->get_status();
+        // defines the statuses when the mail should be send
+        $mailingArray = array(
+            'pending',
+            'on-hold',
+            'processing'
+        );
+
+        if ($this->get_option('send_payment_info') === 'yes') {
+            if (in_array($status, $mailingArray)) {
+                echo $order->get_meta('heidelpay-paymentInfo');
+            }
+        }
     }
 }